added private functions in utils/index.js modified function to connect geth node
